Updating copyright year and holders
Updating the copyright year. Also noticed that the project license had the wrong copyright holder listed. Copyright holders in the license should have been listed as the authors listed in the readme file. The mistake was made due to haste of the DjangoDash competition, which was a 48 hour contest over a weekend. License remains MIT. Enjoy!